Maps

+++ Finder.cpp
@@ -5,7 +5,7 @@
     vector<int> Finder::findSubstrings(string s1, string s2) {
         vector<int> result;
         result.reserve(s2.size());
-        map<string, int> permuations;
+        map<string, size_t> permuations;
 
         for(size_t i=s1.size(); i >= 1; i--){
             for(size_t j=s1.size()-i; j >= 1; j--){
